requirements.txt
certifi==2020.6.20
cffi==1.14.3
cryptography==3.1.1
decorator==4.4.2
emoji==0.6.0
ephem==3.7.7.1
flake8==3.8.3
mccabe==0.6.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PySocks==1.7.1
python-telegram-bot==12.8
six==1.15.0
tornado==6.0.4
